+++ src/api/local.rs
@@ -6,15 +6,50 @@ use crate::{OPTIONS, channel::Video, utils};
 use anyhow::Result;
 use async_trait::async_trait;
 use futures_util::future::join_all;
+use regex_lite::Regex;
 use reqwest::Client;
 use serde_json::Value;
 use std::collections::HashSet;
+use std::sync::{LazyLock, OnceLock};
 use std::time::Duration;
 use std::{io::Write, path::PathBuf};
 
 const API_BACKEND: ApiBackend = ApiBackend::Local;
-const ANDROID_USER_AGENT: &str =
-    "com.google.android.youtube/20.10.38 (Linux; U; Android 12; US) gzip";
+
+enum InnertubeClient {
+    Web,
+    AndroidVR,
+}
+
+impl InnertubeClient {
+    fn get(self) -> Value {
+        match self {
+            InnertubeClient::Web => serde_json::json!({
+                "context": {
+                    "client": {
+                        "clientName": "WEB",
+                        "clientVersion": "2.20260114.08.00"
+                    }
+                }
+            }),
+            InnertubeClient::AndroidVR => serde_json::json!({
+                "context": {
+                    "client": {
+                        "clientName": "ANDROID_VR",
+                        "clientVersion": "1.71.26",
+                        "deviceMake": "Oculus",
+                        "deviceModel": "Quest 3",
+                        "androidSdkVersion": 32,
+                        "userAgent": "com.google.android.apps.youtube.vr.oculus/1.71.26 (Linux; U; Android 12L; eureka-user Build/SQ3A.220605.009.A1) gzip",
+                        "osName": "Android",
+                        "osVersion": "12L",
+                        "visitorData": "CgtLT21YQTlDUjNqbyjMp-jMBjInCgJCRRIhEh0SGwsMDg8QERITFBUWFxgZGhscHR4fICEiIyQlJiAp",
+                    },
+                },
+            }),
+        }
+    }
+}
 
 #[derive(Clone)]
 pub struct Local {
@@ -245,7 +280,6 @@ fn extract_videos_from_tab(tab: &Value) -> Option<&[Value]> {
 impl Local {
     pub fn new() -> Self {
         let client = Client::builder()
-            .user_agent(ANDROID_USER_AGENT)
             .timeout(Duration::from_secs(OPTIONS.request_timeout))
             .build()
             .unwrap();
@@ -258,36 +292,55 @@ impl Local {
         }
     }
 
+    async fn get_visitor_data(&self) -> Result<String> {
+        static VISITOR_DATA: OnceLock<String> = OnceLock::new();
+        static RE: LazyLock<Regex> =
+            LazyLock::new(|| Regex::new(r#""VISITOR_DATA":"(\S*?)""#).unwrap());
+
+        match VISITOR_DATA.get() {
+            Some(data) => Ok(data.to_owned()),
+            None => {
+                let webpage = self
+                    .client
+                    .get("https://www.youtube.com")
+                    .send()
+                    .await?
+                    .text()
+                    .await?;
+
+                let Some(visitor_data) = RE.captures(&webpage).and_then(|c| c.get(1)) else {
+                    return Err(anyhow::anyhow!("Couldn't extract visitor data"));
+                };
+
+                let _ = VISITOR_DATA.set(visitor_data.as_str().to_owned());
+
+                Ok(VISITOR_DATA.get().unwrap().clone())
+            }
+        }
+    }
+
     pub async fn post_player(&self, video_id: &str) -> Result<Value> {
-        let url = "https://www.youtube.com/youtubei/v1/player?key=AIzaSyA8eiZmM1FaDVjRy-df2KTyQ_vz_yYM39w";
+        let url = "https://www.youtube.com/youtubei/v1/player";
 
-        let data = serde_json::json!({
-            "context": {
-                "client": {
-                    "clientName": "ANDROID",
-                    "clientVersion": "20.10.38",
-                    "userAgent": ANDROID_USER_AGENT,
-                },
-            },
-            "videoId": video_id
-        });
+        let mut data = InnertubeClient::AndroidVR.get();
+        let map = data.as_object_mut().unwrap();
+        map.insert(String::from("videoId"), Value::String(video_id.to_owned()));
 
-        let response = self.client.post(url).json(&data).send().await?;
+        let mut request = self.client.post(url);
+
+        if let Ok(visitor_data) = self.get_visitor_data().await {
+            request = request.header("X-Goog-Visitor-Id", visitor_data);
+        }
+
+        let response = request.json(&data).send().await?;
+
         Ok(response.error_for_status()?.json().await?)
     }
 
     pub async fn post_browse(&self, items: &[(&str, &str)]) -> Result<Value> {
         let url = "https://www.youtube.com/youtubei/v1/browse?key=AIzaSyAO_FJ2SlqU8Q4STEHLGCilw_Y9_11qcW8";
 
-        let mut data = serde_json::json!({
-            "context": {
-                "client": {
-                    "clientName": "WEB",
-                    "clientVersion": "2.20240304.00.00"
-                }
-            }
-        });
-
+        let mut data = InnertubeClient::Web.get();
         let map = data.as_object_mut().unwrap();
 
         for (key, value) in items {